+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # jsRules Changelog
 
+## 0.0.4
+
++ Added a simple caching tool (CacheMap)
++ Made rules caching a little more flexible (rules are cached for 15 minutes)
+
+## 0.0.3
+
++ Fixed bug with getting files from the classpath
+
 ## 0.0.2
 
 + Add BETWEEN and NOT_BETWEEN operators